Digital Diary
Using the web concepts we implemented a digital diary in which we can retrieve any information at any time with the database connectivity PHP as backend. The main information is fixing events at any time and it can receive the information through the mails before one day like a Remainder.